Don't return the Comment instance in Post#comments
Post#comments does its work by creating a Comment instance with
cid=null.

It doesn't make sense to return this, and it causes problems
with the runtime call tree mapping we're doing for the WP.com developer
site.

Probably a better way to fix this is to move the `replies` method from
the Comment class to the Post class.