Set default Promise return in absence of callback
Previously, if no callback were passed into the wpcom requests, a
default callback that threw an exception on error was introduced and
executed.

Now, however, we can use the absence of a callback to indicate that we
would prefer to have a promise returned that rejects or resolves
corresponding to the API response.

This requires support for promises, however, meaning that `wpcom.js`
will need to be built inside of a project that transpiles or polyfills.