Handle POST request parameters more intelligently
Since the callback is now optional, this commit adjusts the POST method
inside of the Request object to only rearrange the parameters if a
function was passed.

Previously, the parameters were shuffled around on the assumption that
a callback would be sent either in `body` of `fn`, but this didn't work
when the callback was left out entireley.

Now, the POST function checks if a callback were even passed before
shuffling and it also handles the case where both the `body` and `fn`
are missing.