Settle on a build layout.
After lots of stewing and testing, I have decided to release this in the
following manner:

	- All production files are built into the `dist/` directory
	- The node file is `index.js`
	- The browser file is `wpcom.js`
	- Both files are minified and have source-maps included.
	- The node version depends on the `npm` dependencies
	- The browser version only depends on the ES6 shims, such as Babel's
	  polyfill or `core.js`.

To test the browser version I created an HTML file just like in
`Readme.md` and included the minified `wpcom.js`. First, however, I
introduced a fatal error in the library and rebuilt the release files.
When the Exception appeared in my browser, I was able to click on it in
Chrome's DevTools and see the original source code via the source maps.

I believe that this should be sufficient for our purposes here in that
others are still welcome to use this without node and with the source
map and without requiring a huge file.

	- Minified node build (no libraries included) : 16 KB
	- Minified browser build (with libraries) : 38 KB