Remove prepublish npm script, create publish make target
`prepublish` npm script gets called when npm install is called, which
makes little sense, but is done for backwards compatibility with npm
packages that were not using `prepublish` correctly. See:

https://github.com/npm/npm/issues/3059

So, the general accepted solution is to not use `prepublish` script and
instead create a make target that does the prepublish tasks and then
calls `npm publish`.