Figure out public build process
So, this comes after lots of discussion with @aduth concerning how the
libary should be built as a module.

 - The actual node module imported through `require()` will be a
   Babelified version of the plain source code. This means we will
create a new copy of every file and preserve the directory structure on
`postinstall`. The "main" file will be `dist/index.js`. There will be no
minified bundle for node use.
 - The single include file for browser use will be automatically
   generated on `prepublish` so it should get built every time we
release a new version. That bundle has no dependencies, rather it
bundles and minifies the dependencies into the single file. It loads as
a var, `var WPCOM = ...`